backport PR server start policy tests to 3.3
